StaticPermutations

Base.PermutedDimsArrays.PermutedDimsArray

Core.Tuple

StaticPermutations.AbstractPermutation

StaticPermutations.NoPermutation

StaticPermutations.Permutation

Base.inv

Base.isperm

Base.length

StaticPermutations.append

StaticPermutations.check_permutation

StaticPermutations.identity_permutation

StaticPermutations.isidentity

StaticPermutations.permute

StaticPermutations.prepend

StaticPermutations.relative_permutation

Base.PermutedDimsArrays.PermutedDimsArray — Method

PermutedDimsArray(A, perm::AbstractPermutation) -> B

Alternative PermutedDimsArray constructor taking a static permutation.

In contrast to the base constructors, the returned type is fully inferred here.

Core.Tuple — Method

Tuple(perm::Permutation)

Extract tuple representation of Permutation.

The result can be passed to permutedims and PermutedDimsArray.

Returns nothing if perm is a NoPermutation.

Examples

julia> Tuple(Permutation(3, 2, 1))
(3, 2, 1)

julia> Tuple(NoPermutation()) === nothing
true

StaticPermutations.AbstractPermutation — Type

AbstractPermutation

Abstract type representing a compile-time permutation.

Subtypes are Permutation and NoPermutation.

StaticPermutations.NoPermutation — Type

NoPermutation <: AbstractPermutation

Represents an identity permutation.

It is functionally equivalent to Permutation(1, 2, 3, ...), and is included for convenience.

StaticPermutations.Permutation — Type

Permutation{p} <: AbstractPermutation

Describes a compile-time dimension permutation.

The type parameter p should be a valid permutation such as (3, 1, 2).

Permutation(perm::Vararg{Int})
Permutation(perm::NTuple{N,Int})

Constructs a Permutation.

Example

Both are equivalent:

p1 = Permutation(3, 4)
p2 = Permutation((3, 4))

Base.inv — Method

inv(p::Permutation)
invperm(p::Permutation)

Returns the inverse permutation of p.

Functionally equivalent to Julia's invperm, with the advantage that the result is a compile time constant.

See also relative_permutation.

Examples

julia> p = Permutation(2, 3, 1);

julia> q = inv(p)
Permutation(3, 1, 2)

julia> t_orig = (36, 42, 14);

julia> t_perm = permute(t_orig, p)
(42, 14, 36)

julia> permute(t_perm, q) === t_orig
true

Base.isperm — Method

isperm(perm::AbstractPermutation) -> Bool

Returns true if perm is a valid permutation, false otherwise.

The result is known at compile time.

Examples

julia> isperm(Permutation(3, 1, 2))
true

julia> isperm(Permutation(4, 1, 2))
false

julia> isperm(NoPermutation())
true

Base.length — Method

length(perm::AbstractPermutation)

Returns length of permutation.

For NoPermutation, returns nothing.

Examples

julia> length(Permutation(3, 2, 1))
3

julia> length(NoPermutation()) === nothing
true

StaticPermutations.append — Method

append(p::Permutation, ::Val{M})

Append M non-permuted dimensions to the given permutation.

Examples

julia> append(Permutation(2, 3, 1), Val(2))
Permutation(2, 3, 1, 4, 5)

julia> append(NoPermutation(), Val(2))
NoPermutation()

StaticPermutations.check_permutation — Method

check_permutation(perm::AbstractPermutation)

Check the validity of a Permutation.

Throws ArgumentError if the permutation is invalid.

See also isperm.

Examples

julia> check_permutation(Permutation(3, 2, 1))  # no error

julia> check_permutation(NoPermutation())       # no error

julia> check_permutation(Permutation(3, 3, 1))
ERROR: ArgumentError: invalid permutation of dimensions: Permutation(3, 3, 1)

StaticPermutations.identity_permutation — Method

identity_permutation(::Val{N})
identity_permutation(A::AbstractArray{T,N})

Returns the identity permutation Permutation(1, 2, ..., N).

StaticPermutations.isidentity — Method

isidentity(p::Permutation)

Returns true if p is an identity permutation, i.e. if it is equivalent to (1, 2, 3, ...).

julia> isidentity(Permutation(1, 2, 3))
true

julia> isidentity(Permutation(1, 3, 2))
false

julia> isidentity(NoPermutation())
true

StaticPermutations.permute — Method

permute(collection, perm::Permutation)

Permute collection according to the given permutation.

The collection may be a Tuple, a CartesianIndex, or a Permutation to be reordered according to perm.

Examples

julia> perm = Permutation(2, 3, 1);

julia> permute((36, 42, 14), perm)
(42, 14, 36)

julia> permute(CartesianIndex(36, 42, 14), perm)
CartesianIndex(42, 14, 36)

julia> permute(Permutation(3, 1, 2), perm)
Permutation(1, 2, 3)

StaticPermutations.prepend — Method

prepend(p::Permutation, ::Val{M})

Prepend M non-permuted dimensions to the given permutation.

Examples

julia> prepend(Permutation(2, 3, 1), Val(2))
Permutation(1, 2, 4, 5, 3)

julia> prepend(NoPermutation(), Val(2))
NoPermutation()

StaticPermutations.relative_permutation — Method

relative_permutation(x::Permutation, y::Permutation)

Get relative permutation needed to get from x to y. That is, the permutation perm such that permute(x, perm) == y.

The computation is performed at compile time using generated functions.

Examples

julia> x = Permutation(3, 1, 2);

julia> y = Permutation(2, 1, 3);

julia> perm = relative_permutation(x, y)
Permutation(3, 2, 1)

julia> permute(x, perm) == y
true